backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
sqlparse
requests
pydantic
